Fixed bug in checksum computation during compaction.
During compaction the crc32 of an entry was copied from
the uncompacted version of the entry, which contains information
about the header. As compaction changes the header, this crc32
is incorrect. The crc32 is now recomputed using the key and value,
in order to produce the correct value when the entry is later
persisted to secondary storage.