Objectified the Event Manager.
The Event Manager was using global variables, thus if multiple databases
were used in the same program, they would refer to the same Event Manager
and crash. The Event Manager is now an attribute of the KingDB class,
and passed as a pointer to the Write Buffer and the Storage Engine. It
is not unique to every database, which allows multiple databases to be
opened concurrently.