Added a bandwidth-adaptive write buffer to prevent timeouts for network clients.
- The Writer Buffer was improperly set: it was set to never block and
therefore it would grow without bound and cause Out-Of-Memory crashes.
That was introduced as the previous version was simply blocking when the
buffer was full, and waiting for the Storage Engine to persist the
current batch to secondary storage, but this would cause timeouts for
clients used over the network. Neither solution, either the
never-blocking one or the blocking one are correct. The correct way is
to adapt the bandwitdh of incoming traffic to the bandwidth of the
secondary storage: that way, the writes are as fast as they can be, and
the clients are being slown down just enough to avoid making them
timeout. The 'adaptive' Write Buffer is now the default mode for the
Write Buffer. A 'blocking' mode also be activated through parameters
if needed.
- A rate limiting option has also been added.
- A few parameters were renamed and moved around.
- A unit test for large file was created.