Bug fix in the swapping of the Write Buffer.
The double flushing occurring when the database is being closed or when
a snapshot is being created was not working properly. The reason was
that the same buffer, the 'copy' one, was trying to be flushed on
secondary storage, without swapping it for the 'live' buffer before the
second flush. The two other sources of swapping, a filled live buffer
and a timeout, are still causing swapping as expected.

I fixed the bug, and in the meantime, simplified greatly the way that
the buffer swapping works in the Write Buffer. Swapping is now performed
by only one location of the code instead of two before, and I removed
the 'can_swap_' and 'force_swap_' status variables.