Added support for incompressible data.
Not all data is created equal. Most data will shrink in size when a
compression algorith is applied to it, and some will not, worse, some
data will actually grow in size. I had allocated only enough space to
hold the original data and was thinking this would be enough to store
all data, including incompressible, but I was wrong.

I had to make major changes to the data format to support frames that
hold incompressible data. A lot of the logic for the compression and
checksum are not scattered all around the Storage Engine and the
ByteArray classes, and would require a good clean up at some point.