Fixed a bug in the reverse iteration over the Storage Engine index.
The reverse iteration over ranges from the equal_range() of the
index multimap used to be done by decrementing both iterators
returned by equal_range(). This was causing issues in case one of
these iterators happened to be the first position of the multimap,
i.e. begin(), since decrementing the first position of a container
has an undefined behavior.