Implemented file pool.
The previous implementation was opening a file and making a new mmap for
every single read to the database, which was overkill and of course
impacted performance because of the huge number of system calls. The new
version maintains a pool of file descriptors and mmaps, and just reuse
these as much as possible to avoid opening the same files over and over
again.