Optimized checksum calculation, bringing a 3x speed-up for the writes.
I am so happy this is finally fixed! There used to be only one checksum
field in the EntryHeader, that was guarding both the header and the data
of an entry. But because of the multipart API, there were cases were the
compressed sized of a entry was not known until the very last part was
received, forcing the EntryHeader to be written again. Because computing
the checksum for the data is better done on the fly (so that the data is
not read another time just to compute the checksum), I was doing the
full checksum computation as such:

1. Compute the checksum for the data on the fly, checksum_content.
2. Compute the checksum for the header once all fields are known,
   checksum_header.
3. Call crc32c::Combine() on checksum_content and checksum_header to
   combine both checksums.

When the entry would be later read, the checksum was calculated for the
header and data at the same time, which was possible because of the call
to crc32c::Combine() during the write. But this came with a huge
downside: Combine() is extremely slow. Actually, after doing some
profiling, I found out that when doing writes, 40% of the CPU time was
spent on Combine(). I then decided to fix the problem, by adding a new
field to the EntryHeader, to embed only the checksum for the header
itself in a CRC8, and use the previously used checksum field only for
the entry data (key and value). And this tiny change brought a 3x
speed-up for the writes. This is much more than the 40% of CPU time
recovered, and I believe is due to the fact that the Combine() was just
killing all the caching in the CPU.