Playing with unordered_multimap.
In this branch I am playing with unordered_multimap as a replacement
for multimap with the goal to incrase the performance of random reads.

So far, when running the LevelDB benchmarks, I haven't seen much
improvements. Also, I replaced directly all occurrences of multimap by
unordered_multimap, which created issues as some of the code in the
Compaction() method was expecting the sorting behavior of multimap
to work.

Another thing I want to explore is to have a custom-made hash table or
tree for the index of KingDB.